\name{psyphy-package}
\alias{psyphy-package}
\alias{psyphy}
\docType{package}
\title{
Functions for analyzing psychophysical functions
}
\description{
An assortment of functions that could be useful in analyzing data from pyschophysical experiments. It includes functions for calculating d' from several different experimental designs, links for mafc to be used with the binomial family in glm (and possibly other contexts) and a self-Start function for estimating gamma values for CRT screen calibrations.
}
\details{
\tabular{ll}{
Package: \tab psyphy\cr
Type: \tab Package\cr
Version: \tab 0.0-2\cr
Date: \tab 2007-01-27\cr
License: \tab GPL\cr
}
For the moment, the package contains several functions for calculating
\eqn{d'} for a variety of psychophysical paradigms, some link functions for
the binomial family in glm (and perhaps other functions) for fitting
psychometric functions from mAFC experiments and a self-Start function
for estimating the value of the exponent, gamma, based on the
luminance calibration of the three channels of a CRT-like display.
}
\author{
Kenneth Knoblauch
 <knoblauch@lyon.inserm.frt>
}
\references{
}

\keyword{ package }
\seealso{

}
\examples{
}
